% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputedProvider.R
\name{ComputedProvider}
\alias{ComputedProvider}
\title{Provider for precomputed results}
\description{
Provider for precomputed results

Provider for precomputed results
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_input_limits}{\code{ComputedProvider$get_input_limits()}}
\item \href{#method-compute}{\code{ComputedProvider$compute()}}
\item \href{#method-clone}{\code{ComputedProvider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_input_limits"></a>}}
\if{latex}{\out{\hypertarget{method-get_input_limits}{}}}
\subsection{Method \code{get_input_limits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedProvider$get_input_limits()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute"></a>}}
\if{latex}{\out{\hypertarget{method-compute}{}}}
\subsection{Method \code{compute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedProvider$compute(year, min_wage)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
