% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedProvider.R
\name{DelayedProvider}
\alias{DelayedProvider}
\title{Delayed provider}
\description{
Delayed provider

Delayed provider
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DelayedProvider$new()}}
\item \href{#method-get_input_limits}{\code{DelayedProvider$get_input_limits()}}
\item \href{#method-clone}{\code{DelayedProvider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedProvider$new(delay = 5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_input_limits"></a>}}
\if{latex}{\out{\hypertarget{method-get_input_limits}{}}}
\subsection{Method \code{get_input_limits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedProvider$get_input_limits()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
